pyproject.toml
[project]
name = "abs"
version = "0.1.0"
description = ""
authors = [
    {name = "David Wales",email = "david.wales@swsphn.com.au"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer (>=0.15.2,<0.16.0)",
    "polars (>=1.24.0,<2.0.0)",
    "fastexcel (>=0.13.0,<0.14.0)",
    # Once we drop support for Python 3.10 (no sooner than 2026-10),
    # we can drop this backport dependency.
    "backports.strenum (>=1.3.1,<2.0) ; python_version < '3.11'",
]

[project.scripts]
abs = 'abs.__main__:app'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
